gutman & asaf & tal - did some changes

+++ src/main/java/edu.greenblitz.gblib/encoder/IEncoder.java
@@ -10,7 +10,7 @@ public interface IEncoder {
     /**
      * @return raw amount of ticks so far
      */
-    int getRawTicks();
+    double getRawTicks();
 
     /**
      * @return ticks per second
